added responsive design
