+++ src/App.css
@@ -1,0 +1,15 @@
+input[type="date"]::-webkit-calendar-picker-indicator {
+  filter: invert(1);
+  color: #ffffff;
+  border-radius: 100px;
+  padding: 0.5rem;
+  transition: background-color 1s;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator:hover {
+  background-color: rgb(221, 221, 202);
+  color: black;
+  filter: invert(0);
+  border-radius: 100px;
+  cursor: pointer;
+}
