+++ src/App.js
@@ -1,7 +1,5 @@
 import "./App.css";
-import AadoButton from "./components/AadoButton";
-import HeaderTitle from "./components/HeaderForm/HeaderTitle";
-import HeaderView from "./components/HeaderForm/HeaderView";
+import HeaderView from "./components/Header/HeaderView";
 
 function App() {
   return <HeaderView />;
