+++ src/components/Header/HeaderTitle.jsx
@@ -1,0 +1,14 @@
+function HeaderTitle() {
+  return (
+    <div className="text-white p-7  ">
+      <h1 className="mb-5 sm:text-2xl xl:text-4xl lg:text-4xl md:text-4xl text-xl font-bold">
+        Choose Your Next Events!
+      </h1>
+      <p className="text-sm sm:text-md md:text-md lg:text-lg">
+        Make memories that will last a lifetime. See your favorite events!
+      </p>
+    </div>
+  );
+}
+
+export default HeaderTitle;
