--- src/components/HeaderForm/HeaderTitle.jsx
@@ -1,12 +1,0 @@
-function HeaderTitle() {
-  return (
-    <div className="text-white p-7  ">
-      <h1 className="mb-5 xl:text-4xl lg:text-4xl md:text-4xl sm: text-xl font-bold">
-        Choose Your Next Events!
-      </h1>
-      <p>Make memories that will last a lifetime. See your favorite events!</p>
-    </div>
-  );
-}
-
-export default HeaderTitle;
